refactor: remove deprecated hibernate structures/functions
post rebase fixes

Refs: XRDDEV-2713
